Problem: developing Postgres extensions in C
It can be tedious and error-prone. There's pgrx for Rust and a similar
library for Zig. Both have their pros and cons. They however share the
fact that both languages are fairly significant departures from C and
require additional toolchains.

Solution: use C++ to develop extensions

C++, while not being C (and being subtly incompatible in some areas)
still draws on its most direct relationship with C and is also a
language renowned for efficiency, interesting approach to
meta-programing and some really great libraries.

This snapshot is not a complete library but a first capture of something
marginally useful.