Problem: SRFs don't check for expected types and can return whatever they want
While the documentation says they don't need to pay attention to `expectedDesc`

Solution: check it anyway because SRF execution matches it with `setDesc`

It has to match to or "considered to match". They do care about the number of
attributes and their types, but allow for droppage for as long as the physical
storage matches, so it seems reasonable to enforce this for type safety.