Problem: managing try/catch internal transactions
In tests, we do this whole ceremony with internal subtransactions, and it is tedious.

Solution: wrap it into an `internal_subtransaction` type.

Currently, it does not have direct controls over, it only releases when it goes
out of scope, and you have to determine whether it is going to commit or
rollback at the declaration time. But this can be improved later on.