Problem: tracking memory context for non-by-value types
I made a wrong assumption (having not thought it through) that any pointer
behind a datum would be its own allocation. Of course it is not true. For example,
when results returned from a query done over SPI, the data is allocated in tuples,
not individually.

Solution: allow passing memory_context to from_nullable_datum/from_datum

This allows us to pass a _known_ memory context (if we do know it) to the
constructor of the non-by-value type so that it can track it. We do this for
SPI case because know the context there. In all other context, it'll assume top
memory context, which is never reset.