Problem: can we use custom types?
Solution: support convertible types explicitly

The defining test case is to implement `type_traits`. This required the
introduction of `named_type` type to be able to find the type of
interest. Combined into/from conversion in this trait.

Importantly, handled the case of a conflict with a PFR-derived tuples.
If a type is convertible from nullable datum, and is the only attribute,
it is not considered as a PFR tuple.

Implemented basic syscache lookup.

Fixed up error reporting.