Problem: sometimes functions are called without type info
And that breaks our type checker – it can't make a decision whether
anything's correct.

Solution: if that happens, look this information up in the syscache

There's a little bit of a caveat, it seems, with `_in` functions for
types: the signature might contain just one argument but it seems
Postgres is always specifying that it is calling with three arguments.