Problem: use of threads
While Postgres is (still) single-threaded, we can't properly make use of C++'s
and its ecosystem's multi-threading capabilities. But it would really be
helpful to.

Solution: implement a generic "worker" to handle Postgres work

This worker is intended to work in a single thread and simply processes all
incoming function calls to be executed in that thread.