Problem: unable to retrieve values with their types
When querying for data that we don't (or can't) bind to a pre-determined
type, we can't really type it, and therefore retrieve it.

And that is very limiting as to what should be possible – we should be
able to pass data around without knowing its C++ counterpart type in
advance, at compile time.

Solution: introduce `value` type

This type binds nullable datum together with a type.

In order for this to become feasible, type and datum conversion traits
now have to accommodate for presence of type OID and value (when the
type is determined in runtime).

The API is not particularly beautiful yet (especially `datum_conversion`
receiving an additional argument is not) but it starts paving the way.
Most notable change there is `type_traits` methods becoming instance
methods so we can pass an instance of value to retrieve information in
the runtime when a value is available.

I've only tested a limited number of scenarios related to SPI, but
expect this to grow to support a larger surface.