Problem: spi_executors result don't enforce return types
It seem to have done this accidentally in some tests, but simply
observing that we're iterating over `Args` to get types shows us that
this is wrong.

Solution: ensure we actually iterate the return tuple types instead

This necessitated some signature changes (dropping `Args` is obvious)
and promotion of `cppgres::oid` into a struct so we can implement
correct conversion and type traits on it.