Problem: use of string_view in executor queries
Because we take pointer to what's in string_view, we actually pass
much more to Postgres than we should:

```
select trueSOMETHING IS HERE
```

Even though the span might be set correctly, it won't be passed to Postgres
correctly and the entire string will be visible.

Solution: convert into std::string internally before getting a pointer

We should likely rethink the type signature for these methods but I wanted to
make something working quickly.