Problem: call info disjoining arg datums and arg types
In `function_call_info` it is possible either to get arg datum iterator or arg
type iterator, but not both.

While this is not a problem in C++23 (due to zip views), it's extra work and is
not readily available in C++20.

Solution: use `value` type to represent the joined argument
and expose it through `arg_values()`.