Problem: writing aggregates in C++
It's possible but involves a lot of manual work. The promise of this library,
however, is to make things trivial.

Solution: design a conceptual API for it

It introduces a concepts of aggregate, finalizable aggregate and more to come.
At this time, just the simple case is handled as a proof of concept.

This is a relatively large change that introduces a change to
`datum_conversion` trait, adding `from_nullable_datum`, and introducing
`default_datum_conversion` that removes some of the boilerplate. This is a good
change as it vastly simplified `from_nullable_datum`, removing a lot of
special-casing.