Problem: writing aggregates in C++
It's possible but involves a lot of manual work. The promise of this library,
however, is to make things trivial.

Solution: design a conceptual API for it

It still misses a few things, like moving-aggregate mode, final func extra and so on.

This is a relatively large addition that introduces a change to
`datum_conversion` trait, adding `from_nullable_datum`, and introducing
`default_datum_conversion` that removes some of the boilerplate. This is a good
change as it vastly simplified `from_nullable_datum`, removing a lot of
special-casing.