Problem: using rendezvous variables
The API for them is trivial, but it is perhaps a bit deceptively simple.

One of the biggest problems is what if we need to access information
provided by extension A but it is loaded after our extension is loaded?

The other one is that `find_rendezvous_variable` gives you a pointer to
a pointer but there's still a chance that the process will crash before
the memory is allocated or initialized? It will leave the rendezvous
variable in an inconsistent state.

Solution: introduce a name registry to better support the workflows

We are not borrowing the same name ("rendezvous") because:

1. it is somewhat hard to spell and is not the most common term
2. can be confusing as we are building a bit more, and term overload is
   not great

The registry allows for `noexcept` initialization of values and,
importantly, an ability to attach a callback for when the name does get
initialized.