Problem: complex node visitors
Right now, we can simply dispatch nodes by type, which is very
convenient to avoid large case switches.

Solution: allow for more complex handling, including recursion

This leverages existing walkers to absorb all the know-how. In the
future, we may review this.

There are some deficiencies in the current implementation:

* We produce some template bloat
* We don't integrate with parsing and processing API (and I think we should)
* The API is a bit clunky

From the original implementation, we still inherit a certain level of
inefficiencies in the dispatch – we still have to call `is` method to
check if the target type is of a certain kind, but we need to have a
better form of direct dispatch, ideally. Though it would be interesting
to see how much of a real impact can this make in practical terms. Still
important to acknoweldge that this is slower than a switch statement and
more advanced techniques like computed goto labels.