Problem: calling Postgres functions is somewhat painful
One needs to call one of the C functions for a specific arity, convert
arguments to and from datums and ensure that these functions are
actually expecting these arguments.

Solution: wrap it all into a cppgres::function struct

It's a bit limited (no collation support at the moment) and I can see
how it can grow further into these directions:

* direct function call (C)
* calling with tuples, arrays and vectors of arguments for dynamic cases

But it's a start.

Had to fix some mistakes and ommissions in optional <-> datum conversion
(which is important!)