Problem: transforming errors into exceptions
By itself, not a bad idea. This way, the C++ layer can catch them or
let them pass through and re-surface as errors.

However, if the process was holding an `LWLock` before, the
`InterruptHoldOffCount` would be incremented for each lock, but
`errfinish` would just set it to zero on `ERROR`. And when `LWLockRelease`
is called, it is asserting that this count should be above zero,
resulting in an undefined behavior, or a trap in assertion-enabled mode.

Solution: restore `InterruptHoldOffCount` upon error

`errfinish` says this:

```
We do some minimal cleanup before longjmp'ing so that handlers can
execute in a reasonably sane state.

Reset InterruptHoldoffCount in case we ereport'd from inside an
interrupt holdoff section.  (We assume here that no handler will
itself be inside a holdoff section.  If necessary, such a handler
could **save and restore** InterruptHoldoffCount for itself, but this
should make life easier for most.)
```

(empahsis mine)

Following this logic, this allows us to keep the count correct.